Module Calculon_redis_lib__Message_j

type privmsg = Calculon_redis_lib.Message_t.privmsg = {

source : string;

dest : string;

message : string;

}

type message = Calculon_redis_lib.Message_t.message

type send = Calculon_redis_lib.Message_t.send = {

message : message;

}

val write_privmsg : Bi_outbuf.t -> privmsg -> unit

Output a JSON value of type privmsg.

val string_of_privmsg : ?⁠len:int -> privmsg -> string

Serialize a value of type privmsg into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_privmsg : Yojson.Safe.lexer_state -> Lexing.lexbuf -> privmsg

Input JSON data of type privmsg.

val privmsg_of_string : string -> privmsg

Deserialize JSON data of type privmsg.

val write_message : Bi_outbuf.t -> message -> unit

Output a JSON value of type message.

val string_of_message : ?⁠len:int -> message -> string

Serialize a value of type message into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_message : Yojson.Safe.lexer_state -> Lexing.lexbuf -> message

Input JSON data of type message.

val message_of_string : string -> message

Deserialize JSON data of type message.

val write_send : Bi_outbuf.t -> send -> unit

Output a JSON value of type send.

val string_of_send : ?⁠len:int -> send -> string

Serialize a value of type send into a JSON string.

parameter len

specifies the initial length of the buffer used internally. Default: 1024.

val read_send : Yojson.Safe.lexer_state -> Lexing.lexbuf -> send

Input JSON data of type send.

val send_of_string : string -> send

Deserialize JSON data of type send.